requirements.txt
requests==2.31.0
pytest==7.4.3
allure-pytest==2.13.2
jsonschema==4.20.0
urllib3==2.1.0
loguru==0.7.2

